Renamed cirq test files
